% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacciSequence}
\alias{fibonacciSequence}
\title{Compute fibonacci sequence}
\usage{
fibonacciSequence(x)
}
\description{
This function computes fibonacci sequence.
}
\examples{

fibonacciSequence(10)

}
