% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fibonacci.R
\name{yourfunction}
\alias{yourfunction}
\title{Fibonacci Sequence}
\usage{
yourfunction(a)
}
\description{
Fibonacci sequence is an array of numbers in which the sum of the first two numbers is immediately followed.
Enter n to indicate the value of the nth Fibonacci sequence.
}
\examples{

FibonacciSequence(8)
}
